img-1128 - problem related to int8 storage of labels. Convert to int before doing math.
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CellProfiler@10456 1fc53939-2000-0410-845c-e8453a809027

+++ cellprofiler/modules/trackobjects.py
@@ -1473,7 +1473,9 @@ class TrackObjects(cpm.CPModule):
                 score to assign new label numbers so that the new objects
                 that are "better" inherit the old objects' label number
         '''
-        old_object_numbers = self.get_saved_object_numbers(workspace)
+        new_of_old = new_of_old.astype(int)
+        old_of_new = old_of_new.astype(int)
+        old_object_numbers = self.get_saved_object_numbers(workspace).astype(int)
         max_object_number = self.get_max_object_number(workspace)
         old_count = len(new_of_old)
         new_count = len(old_of_new)
